terminar ejercicio 3 de prog

+++ Programacion/ExtraT3/src/ejercicio03/Vendedor.java
@@ -3,5 +3,5 @@ package ejercicio03;
 public class Vendedor {
 	
 	private Movil [] listaMoviles;
-	private int totalVendido;
+	private double totalVendido;
 }
